+++ delphi-frontend/src/main/java/org/sonar/plugins/communitydelphi/api/ast/PropertyImplementsSpecifierNode.java
@@ -1,6 +1,6 @@
 /*
  * Sonar Delphi Plugin
- * Copyright (C) 2023 Integrated Application Development
+ * Copyright (C) 2024 Integrated Application Development
  *
  * This program is free software; you can redistribute it and/or
  * modify it under the terms of the GNU Lesser General Public
@@ -16,30 +16,10 @@
  * License along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02
  */
-package org.sonar.plugins.communitydelphi.api.symbol.declaration;
+package org.sonar.plugins.communitydelphi.api.ast;
 
 import java.util.List;
-import javax.annotation.Nullable;
-import org.sonar.plugins.communitydelphi.api.ast.Visibility;
-import org.sonar.plugins.communitydelphi.api.symbol.Invocable;
-import org.sonar.plugins.communitydelphi.api.type.Type;
 
-public interface PropertyNameDeclaration extends TypedDeclaration, Invocable, Visibility {
-  String fullyQualifiedName();
-
-  @Nullable
-  NameDeclaration getReadDeclaration();
-
-  @Nullable
-  NameDeclaration getWriteDeclaration();
-
-  List<Type> getImplementedTypes();
-
-  boolean isArrayProperty();
-
-  boolean isDefaultProperty();
-
-  List<PropertyNameDeclaration> getRedeclarations();
-
-  List<Type> getAttributeTypes();
+public interface PropertyImplementsSpecifierNode extends DelphiNode {
+  List<TypeReferenceNode> getTypeReferences();
 }