+++ delphi-frontend/src/test/resources/au/com/integradev/delphi/symbol/properties/ImplementsProperties.pas
@@ -1,0 +1,18 @@
+unit ImplementsProperties;
+
+interface
+
+type
+  IBar = interface
+  end;
+
+  TBar = class(IBar)
+  end;
+
+  TFoo = class
+    property Bar: TBar implements IBar;
+  end;
+
+implementation
+
+end.