+++ delphi-frontend/src/test/resources/au/com/integradev/delphi/symbol/properties/IndexProperties.pas
@@ -1,0 +1,15 @@
+unit IndexProperties;
+
+interface
+
+const
+  C_Bar = 123;
+
+type
+  TFoo = class
+    property Bar: Integer index C_Bar;
+  end;
+
+implementation
+
+end.